% Generated by roxygen2: do not edit by hand
% Please edit documentation in utilities/dEploidTools.r
\name{computeObsWSAF}
\alias{computeObsWSAF}
\title{Compute observed WSAF}
\usage{
computeObsWSAF(alt, ref)
}
\arguments{
\item{alt}{Numeric array of alternative allele count.}

\item{ref}{Numeric array of reference allele count.}
}
\value{
Numeric array of observed allele frequency within sample.
}
\description{
Compute observed allele frequency within sample from the allele counts.
}
\examples{
vcfFile = system.file("extdata", "PG0390-C.test.vcf.gz", package = "DEploid")
PG0390 = extractCoverageFromVcf(vcfFile)
computeObsWSAF( PG0390$altCount, PG0390$refCount )

}
\seealso{
\code{\link{histWSAF}} for histogram, \code{\link{plotWSAF}} for WSAFs plotted against indices.
}

