\name{DEploid-package}
\alias{DEploid-package}
\alias{DEploid}
\docType{package}
\title{
\packageTitle{DEploid}
}
\description{
\packageDescription{DEploid}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DEploid}
\packageIndices{DEploid}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{DEploid}

Maintainer: \packageMaintainer{DEploid}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
