\name{DEploid-r-package}
\alias{DEploid-r-package}
\alias{DEploid-r}
\docType{package}
\title{
\packageTitle{DEploid-r}
}
\description{
\packageDescription{DEploid-r}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DEploid-r}
\packageIndices{DEploid-r}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{DEploid-r}

Maintainer: \packageMaintainer{DEploid-r}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
