% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dEploidTools.R
\name{haplotypePainter}
\alias{haplotypePainter}
\title{Painting haplotype according the reference panel}
\usage{
haplotypePainter(posteriorProbabilities, title = "")
}
\arguments{
\item{posteriorProbabilities}{Posterior probabilities matrix with the size of number of loci by the number of reference strain.}

\item{title}{Figure title.}
}
\description{
Plot the posterior probabilities of a haplotype given the refernece panel.
}

