% Generated by roxygen2: do not edit by hand
% Please edit documentation in utilities/dEploidTools.r
\name{plot.prob}
\alias{plot.prob}
\title{Deconvolute mixed haplotypes, and reporting the mixture proportions from each sample
This function provieds an interface for calling \emph{dEploid} from R.
The command line options are passed via the \code{args} argument and \code{file}}
\usage{
\method{plot}{prob}(tmpProp, title)
}
\arguments{
\item{tmpProp}{blah blah blah}

\item{title}{blah blah blah}
}
\value{
A named list of something something ...
}
\description{
Deconvolute mixed haplotypes, and reporting the mixture proportions from each sample
This function provieds an interface for calling \emph{dEploid} from R.
The command line options are passed via the \code{args} argument and \code{file}
}
\section{blahblah}{

Blah blah
}
\examples{
set.seed(1234)

}

