% Generated by roxygen2: do not edit by hand
% Please edit documentation in utilities/dEploidTools.r
\name{plotWSAFvsPLAF}
\alias{plotWSAFvsPLAF}
\title{Plot WSAF vs PLAF}
\usage{
plotWSAFvsPLAF(plaf, obsWSAF, expWSAF = c(), title = "WSAF vs PLAF")
}
\arguments{
\item{plaf}{Numeric array of population level allele frequency.}

\item{obsWSAF}{Numeric array of observed altenative allele frequencies within sample.}

\item{expWSAF}{Numeric array of expected WSAF from model.}

\item{title}{Figure title, "WSAF vs PLAF" by default}
}
\value{

}
\description{
Plot allele frequencies within sample against population level.
}
\examples{
vcfFile = system.file("extdata", "PG0390-C.test.vcf.gz", package = "DEploid")
PG0390 = extractCoverageFromVcf(vcfFile)
obsWSAF = computeObsWSAF( PG0390$altCount, PG0390$refCount )
plafFile = system.file("extdata", "labStrains.test.PLAF.txt", package = "DEploid")
plaf = extractPLAF(plafFile)
plotWSAFvsPLAF(plaf, obsWSAF)

}

