% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploidPlotly.R
\name{plotWSAFVsPLAFplotly}
\alias{plotWSAFVsPLAFplotly}
\title{Plot WSAF vs PLAF}
\usage{
plotWSAFVsPLAFplotly(plaf, obsWSAF, ref, alt)
}
\arguments{
\item{plaf}{Numeric array of population level allele frequency.}

\item{obsWSAF}{Numeric array of observed altenative allele frequencies within
sample.}

\item{ref}{Numeric array of reference allele count.}

\item{alt}{Numeric array of alternative allele count.}
}
\description{
Plot allele frequencies within sample against population level.
}
\examples{
# Example 1
refFile = system.file("extdata", "PG0390-C.test.ref", package = "DEploid")
altFile = system.file("extdata", "PG0390-C.test.alt", package = "DEploid")
PG0390CoverageT = extractCoverageFromTxt(refFile, altFile)
obsWSAF = computeObsWSAF(PG0390CoverageT$altCount, PG0390CoverageT$refCount)
plafFile = system.file("extdata", "labStrains.test.PLAF.txt",
  package = "DEploid")
plaf = extractPLAF(plafFile)
plotWSAFvsPLAF(plaf, obsWSAF, PG0390CoverageT$refCount,
               PG0390CoverageT$altCount)

# Example 2
vcfFile = system.file("extdata", "PG0390-C.test.vcf.gz", package = "DEploid")
PG0390CoverageV = extractCoverageFromVcf(vcfFile)
obsWSAF = computeObsWSAF(PG0390CoverageV$altCount, PG0390CoverageV$refCount)
plafFile = system.file("extdata", "labStrains.test.PLAF.txt",
  package = "DEploid")
plaf = extractPLAF(plafFile)
plotWSAFvsPLAF(plaf, obsWSAF, PG0390CoverageV$refCount,
               PG0390CoverageV$altCount)

}
