% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_hello_world}
\alias{rcpp_hello_world}
\title{sasfdafdsf asdfsdafa}
\usage{
rcpp_hello_world()
}
\value{
A named list of something something ...
}
\description{
sasfdafdsf asdfsdafa
}
\section{blahblah}{

Blah blah
}
\examples{
set.seed(1234)

}
\keyword{datagen}

