% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dEploid}
\alias{dEploid}
\title{Deconvolute mixed haplotypes}
\usage{
dEploid(args)
}
\arguments{
\item{args}{String of dEploid input.}
}
\value{
A list with members of haplotypes, proportions and log likelihood of the MCMC chain.
\itemize{
  \item \code{Haps} Haplotypes at the final iteration in plain text file.
  \item \code{Proportions} MCMC updates of the proportion estimates.
  \item \code{llks} Log likelihood of the MCMC chain.
}
}
\description{
Deconvolute mixed haplotypes, and reporting the mixture proportions from each sample
This function provieds an interface for calling \emph{dEploid} from R.
The command line options are passed via the \code{args} argument
}
\section{Seeding}{

The R version of DEploid uses random number from R's random generator. Therefore, the '-seed'
argument of the command line version will be ignored, and no seed is given in the output. Use
the R function ‘set.seed’ prior to calling this function to ensure reproduciblity of results.
}
\examples{
# Explore the data
vcfFile = system.file("extdata", "PG0390-C.test.vcf.gz", package = "DEploid")
PG0390 = extractCoverageFromVcf(vcfFile)
par(mfrow=c(1,3))
plotAltVsRef( PG0390$refCount, PG0390$altCount )
obsWSAF = computeObsWSAF( PG0390$altCount, PG0390$refCount )
histWSAF(obsWSAF)

# Load prior information: PLAF and reference panel
plafFile = system.file("extdata", "labStrains.test.PLAF.txt", package = "DEploid")
plaf = extractPLAF(plafFile)
plotWSAFvsPLAF(plaf, obsWSAF)
panelFile = system.file("extdata", "labStrains.test.panel.txt", package = "DEploid")

# Deconvolute the haplotypes
set.seed(1234)
PG0390.deconv = dEploid(paste("-vcf", vcfFile, "-plaf", plafFile, "-panel", panelFile))
par(mfrow = c(1,3))
prop = PG0390.deconv$Proportions[dim(PG0390.deconv$Proportions)[1],]
expWSAF = t(PG0390.deconv$Haps) \%*\% prop
plotWSAFvsPLAF(plaf, obsWSAF, expWSAF)
plotProportions(PG0390.deconv$Proportions)
plotObsExpWSAF(obsWSAF, expWSAF)

}
\seealso{
\itemize{
  \item \code{vignette('dEploid-Arguments')} for an overview of commandline arguments
}
}

