% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dEploid}
\alias{dEploid}
\title{Deconvolute mixed haplotypes, and reporting the mixture proportions from each sample
This function provieds an interface for calling \emph{dEploid} from R.
The command line options are passed via the \code{args} argument.}
\usage{
dEploid(args, file = "")
}
\arguments{
\item{args}{blah blah blah}

\item{file}{blah blah blah}
}
\value{
A named list of something something ...
}
\description{
Deconvolute mixed haplotypes, and reporting the mixture proportions from each sample
This function provieds an interface for calling \emph{dEploid} from R.
The command line options are passed via the \code{args} argument.
}
\section{blahblah}{

Blah blah
}
\examples{
set.seed(1234)


}
\keyword{datagen}

