% Generated by roxygen2: do not edit by hand
% Please edit documentation in utilities/dEploidTools.r
\name{histWSAF}
\alias{histWSAF}
\title{WSAF histogram}
\usage{
histWSAF(obsWSAF, exclusive = TRUE, title = "Histogram 0<WSAF<1")
}
\arguments{
\item{obsWSAF}{Observed allele frequency within sample}

\item{exclusive}{When TRUE 0 < WSAF < 1; otherwise 0 <= WSAF <= 1.}

\item{title}{Histogram title}
}
\value{
histogram
}
\description{
Produce histogram of the allele frequency within sample.
}
\examples{
vcfFile = system.file("extdata", "PG0390-C.test.vcf.gz", package = "DEploid")
PG0390 = extractCoverageFromVcf(vcfFile)
obsWSAF = computeObsWSAF( PG0390$altCount, PG0390$refCount )
histWSAF(obsWSAF)
myhist = histWSAF(obsWSAF, FALSE)

}

