% Generated by roxygen2: do not edit by hand
% Please edit documentation in utilities/dEploidTools.r
\name{plotAltVsRef}
\alias{plotAltVsRef}
\title{Plot coverage}
\usage{
plotAltVsRef(ref, alt, title = "Alt vs Ref", exclude.ref = c(),
  exclude.alt = c())
}
\arguments{
\item{ref}{Numeric array of reference allele count.}

\item{alt}{Numeric array of alternative allele count.}

\item{title}{Figure title, "Alt vs Ref" by default}

\item{exclude.ref}{Numeric array of reference allele count at sites that are not deconvoluted.}

\item{exclude.alt}{Numeric array of alternative allele count at sites that are not deconvoluted}
}
\description{
Plot alternative allele count vs reference allele count at each site.
}
\examples{
vcfFile = system.file("extdata", "PG0390-C.test.vcf.gz", package = "DEploid")
PG0390 = extractCoverageFromVcf(vcfFile)
plotAltVsRef( PG0390$refCount, PG0390$altCount )

}

