% Generated by roxygen2: do not edit by hand
% Please edit documentation in utilities/dEploidTools.r
\name{plotObsExpWSAF}
\alias{plotObsExpWSAF}
\title{Plot WSAF}
\usage{
plotObsExpWSAF(obsWSAF, expWSAF, title = "WSAF(observed vs expected)")
}
\arguments{
\item{obsWSAF}{Numeric array of observed WSAF.}

\item{expWSAF}{Numeric array of expected WSAF.}

\item{title}{Figure title.}
}
\description{
Plot observed alternative allele frequency within sample against expected WSAF.
}
\examples{
vcfFile = system.file("extdata", "PG0390-C.test.vcf.gz", package = "DEploid")
PG0390 = extractCoverageFromVcf(vcfFile)
obsWSAF = computeObsWSAF( PG0390$altCount, PG0390$refCount )
PG0390.deconv = dEploid(paste("-vcf", vcfFile, "-plaf", plafFile, "-noPanel"))
prop = PG0390.deconv$Proportions[dim(PG0390.deconv$Proportions)[1],]
expWSAF = t(PG0390.deconv$Haps) \%*\% prop
plotObsExpWSAF(obsWSAF, expWSAF)

}

