% Generated by roxygen2: do not edit by hand
% Please edit documentation in utilities/dEploidTools.r
\name{plotProportions}
\alias{plotProportions}
\title{Plot proportions}
\usage{
plotProportions(proportions, title = "Components")
}
\arguments{
\item{proportions}{Matrix of the MCMC proportion samples. The matrix size is number of the MCMC samples by the number of strains.}

\item{title}{Figure title.}
}
\description{
Plot the MCMC samples of the proportion, indexed by the MCMC chain.
}
\examples{
plafFile = system.file("extdata", "labStrains.test.PLAF.txt", package = "DEploid")
vcfFile = system.file("extdata", "PG0390-C.test.vcf.gz", package = "DEploid")
panelFile = system.file("extdata", "labStrains.test.panel.txt", package = "DEploid")
PG0390 = dEploid(paste("-vcf", vcfFile, "-plaf", plafFile, "-noPanel"))
plotProportions( PG0390$Proportions, "PG0390-C proportions" )

}

