% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dEploidTools.R
\name{plotWSAFvsPLAF}
\alias{plotWSAFvsPLAF}
\title{Plot WSAF vs PLAF}
\usage{
plotWSAFvsPLAF(plaf, obsWSAF, expWSAF = c(), title = "WSAF vs PLAF")
}
\arguments{
\item{plaf}{Numeric array of population level allele frequency.}

\item{obsWSAF}{Numeric array of observed altenative allele frequencies within sample.}

\item{expWSAF}{Numeric array of expected WSAF from model.}

\item{title}{Figure title, "WSAF vs PLAF" by default}
}
\description{
Plot allele frequencies within sample against population level.
}
\examples{
# Example 1
refFile = system.file("extdata", "PG0390-C.test.ref", package = "DEploid")
altFile = system.file("extdata", "PG0390-C.test.alt", package = "DEploid")
PG0390CoverageTxt = extractCoverageFromTxt(refFile, altFile)
obsWSAF = computeObsWSAF( PG0390CoverageTxt$altCount, PG0390CoverageTxt$refCount )
plafFile = system.file("extdata", "labStrains.test.PLAF.txt", package = "DEploid")
plaf = extractPLAF(plafFile)
plotWSAFvsPLAF(plaf, obsWSAF)

# Example 2
vcfFile = system.file("extdata", "PG0390-C.test.vcf.gz", package = "DEploid")
PG0390CoverageVcf = extractCoverageFromVcf(vcfFile)
obsWSAF = computeObsWSAF( PG0390CoverageVcf$altCount, PG0390CoverageVcf$refCount )
plafFile = system.file("extdata", "labStrains.test.PLAF.txt", package = "DEploid")
plaf = extractPLAF(plafFile)
plotWSAFvsPLAF(plaf, obsWSAF)

}

